services/schwab-automation/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "schwab-automation"
version = "0.1.0"
description = "Hourly Schwab automation worker for account reconciliation and trade management"
requires-python = ">=3.11"
dependencies = [
  "github-copilot-sdk>=0.1.32",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "schwab-py>=1.5.0",
  "swing-agent-database",
]

[project.scripts]
schwab-automation = "schwab_automation.main:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
schwab_automation = ["prompts/*.md"]